moved transform u2i/i2i tests to transformer_data_preparator
