+++ tests/models/nn/test_bert4rec.py
@@ -7,12 +7,11 @@ import torch
 from pytorch_lightning import Trainer, seed_everything
 
 from rectools.columns import Columns
-from rectools.dataset import Dataset, IdMap, Interactions
+from rectools.dataset import Dataset
 from rectools.models import BERT4RecModel
 from rectools.models.nn.bert4rec import MASKING_VALUE, PADDING_VALUE, BERT4RecDataPreparator
 from rectools.models.nn.item_net import IdEmbeddingsItemNet
 from tests.models.utils import assert_second_fit_refits_model
-from tests.testing_utils import assert_id_map_equal, assert_interactions_set_equal
 
 
 class TestBERT4RecModel:
@@ -472,120 +471,6 @@ class TestBERT4RecDataPreparator:
         )
 
     @pytest.mark.parametrize(
-        "expected_user_id_map, expected_item_id_map, expected_interactions",
-        (
-            (
-                IdMap.from_values([30, 40, 10]),
-                IdMap.from_values(["PAD", "MASK", 15, 11, 12, 17, 14, 13]),
-                Interactions(
-                    pd.DataFrame(
-                        [
-                            [0, 2, 1.0, "2021-11-25"],
-                            [1, 3, 1.0, "2021-11-25"],
-                            [0, 4, 2.0, "2021-11-26"],
-                            [1, 5, 1.0, "2021-11-26"],
-                            [0, 3, 1.0, "2021-11-27"],
-                            [2, 6, 1.0, "2021-11-28"],
-                            [2, 3, 1.0, "2021-11-29"],
-                            [2, 4, 1.0, "2021-11-29"],
-                            [2, 7, 1.0, "2021-11-30"],
-                        ],
-                        columns=[Columns.User, Columns.Item, Columns.Weight, Columns.Datetime],
-                    ),
-                ),
-            ),
-        ),
-    )
-    def test_process_dataset_train(
-        self,
-        dataset: Dataset,
-        data_preparator: BERT4RecDataPreparator,
-        expected_interactions: Interactions,
-        expected_item_id_map: IdMap,
-        expected_user_id_map: IdMap,
-    ) -> None:
-        actual = data_preparator.process_dataset_train(dataset)
-        assert_id_map_equal(actual.user_id_map, expected_user_id_map)
-        assert_id_map_equal(actual.item_id_map, expected_item_id_map)
-        assert_interactions_set_equal(actual.interactions, expected_interactions)
-
-    @pytest.mark.parametrize(
-        "expected_user_id_map, expected_item_id_map, expected_interactions",
-        (
-            (
-                IdMap.from_values([10, 20]),
-                IdMap.from_values(["PAD", "MASK", 15, 11, 12, 17, 14, 13]),
-                Interactions(
-                    pd.DataFrame(
-                        [
-                            [0, 7, 1.0, "2021-11-30"],
-                            [0, 3, 1.0, "2021-11-29"],
-                            [0, 4, 1.0, "2021-11-29"],
-                            [0, 6, 1.0, "2021-11-28"],
-                            [1, 7, 9.0, "2021-11-28"],
-                        ],
-                        columns=[Columns.User, Columns.Item, Columns.Weight, Columns.Datetime],
-                    ),
-                ),
-            ),
-        ),
-    )
-    def test_transform_dataset_u2i(
-        self,
-        dataset: Dataset,
-        data_preparator: BERT4RecDataPreparator,
-        expected_interactions: Interactions,
-        expected_item_id_map: IdMap,
-        expected_user_id_map: IdMap,
-    ) -> None:
-        data_preparator.process_dataset_train(dataset)
-        users = [10, 20]
-        actual = data_preparator.transform_dataset_u2i(dataset, users)
-        assert_id_map_equal(actual.user_id_map, expected_user_id_map)
-        assert_id_map_equal(actual.item_id_map, expected_item_id_map)
-        assert_interactions_set_equal(actual.interactions, expected_interactions)
-
-    @pytest.mark.parametrize(
-        "expected_user_id_map, expected_item_id_map, expected_interactions",
-        (
-            (
-                IdMap.from_values([10, 30, 40, 50, 20]),
-                IdMap.from_values(["PAD", "MASK", 15, 11, 12, 17, 14, 13]),
-                Interactions(
-                    pd.DataFrame(
-                        [
-                            [0, 7, 1.0, "2021-11-30"],
-                            [0, 3, 1.0, "2021-11-29"],
-                            [0, 4, 1.0, "2021-11-29"],
-                            [1, 3, 1.0, "2021-11-27"],
-                            [1, 4, 2.0, "2021-11-26"],
-                            [1, 2, 1.0, "2021-11-25"],
-                            [2, 3, 1.0, "2021-11-25"],
-                            [2, 5, 1.0, "2021-11-26"],
-                            [0, 6, 1.0, "2021-11-28"],
-                            [4, 7, 9.0, "2021-11-28"],
-                        ],
-                        columns=[Columns.User, Columns.Item, Columns.Weight, Columns.Datetime],
-                    ),
-                ),
-            ),
-        ),
-    )
-    def test_tranform_dataset_i2i(
-        self,
-        dataset: Dataset,
-        data_preparator: BERT4RecDataPreparator,
-        expected_interactions: Interactions,
-        expected_item_id_map: IdMap,
-        expected_user_id_map: IdMap,
-    ) -> None:
-        data_preparator.process_dataset_train(dataset)
-        actual = data_preparator.transform_dataset_i2i(dataset)
-        assert_id_map_equal(actual.user_id_map, expected_user_id_map)
-        assert_id_map_equal(actual.item_id_map, expected_item_id_map)
-        assert_interactions_set_equal(actual.interactions, expected_interactions)
-
-    @pytest.mark.parametrize(
         "train_batch",
         (
             (
@@ -619,7 +504,7 @@ class TestBERT4RecDataPreparator:
             ),
         ),
     )
-    def test_get_dataloader_train_masked_session_with_random_replacement(
+    def test_get_dataloader_train_for_masked_session_with_random_replacement(
         self, dataset_one_session: Dataset, train_batch: tp.List
     ) -> None:
         data_preparator = BERT4RecDataPreparator(
