+++ tests/models/nn/test_transformer_data_preparator.py
@@ -16,9 +16,16 @@ import typing as tp
 
 import pandas as pd
 import pytest
+import torch
+from pytorch_lightning import seed_everything
 
 from rectools.columns import Columns
-from rectools.models.nn.transformer_data_preparator import SequenceDataset
+from rectools.dataset import Dataset, IdMap, Interactions
+from rectools.models.nn.sasrec import PADDING_VALUE
+from rectools.models.nn.transformer_data_preparator import SequenceDataset, SessionEncoderDataPreparatorBase
+from tests.testing_utils import assert_id_map_equal, assert_interactions_set_equal
+
+from ..data import INTERACTIONS
 
 
 class TestSequenceDataset:
@@ -59,3 +66,186 @@ class TestSequenceDataset:
         )
         assert len(actual.weights) == len(expected_weights)
         assert all(actual_list == expected_list for actual_list, expected_list in zip(actual.weights, expected_weights))
+
+
+class TestSessionEncoderDataPreparatorBase:
+
+    def setup_method(self) -> None:
+        self._seed_everything()
+
+    def _seed_everything(self) -> None:
+        torch.use_deterministic_algorithms(True)
+        seed_everything(32, workers=True)
+
+    @pytest.fixture
+    def dataset(self) -> Dataset:
+        interactions_df = pd.DataFrame(
+            [
+                [10, 13, 1, "2021-11-30"],
+                [10, 11, 1, "2021-11-29"],
+                [10, 12, 1, "2021-11-29"],
+                [30, 11, 1, "2021-11-27"],
+                [30, 12, 2, "2021-11-26"],
+                [30, 15, 1, "2021-11-25"],
+                [40, 11, 1, "2021-11-25"],
+                [40, 17, 1, "2021-11-26"],
+                [50, 16, 1, "2021-11-25"],
+                [10, 14, 1, "2021-11-28"],
+                [10, 16, 1, "2021-11-27"],
+                [20, 13, 9, "2021-11-28"],
+            ],
+            columns=Columns.Interactions,
+        )
+        return Dataset.construct(interactions_df)
+
+    @pytest.fixture
+    def dataset_dense_item_features(self) -> Dataset:
+        item_features = pd.DataFrame(
+            [
+                [11, 1, 1],
+                [12, 1, 2],
+                [13, 1, 3],
+                [14, 2, 1],
+                [15, 2, 2],
+                [17, 2, 3],
+            ],
+            columns=[Columns.Item, "f1", "f2"],
+        )
+        ds = Dataset.construct(
+            INTERACTIONS,
+            item_features_df=item_features,
+            cat_item_features=["f1", "f2"],
+            make_dense_item_features=True,
+        )
+        return ds
+
+    @pytest.fixture
+    def data_preparator(self) -> SessionEncoderDataPreparatorBase:
+        return SessionEncoderDataPreparatorBase(
+            session_max_len=3,
+            batch_size=4,
+            dataloader_num_workers=0,
+            item_extra_tokens=(PADDING_VALUE,),
+        )
+
+    @pytest.mark.parametrize(
+        "expected_user_id_map, expected_item_id_map, expected_interactions",
+        (
+            (
+                IdMap.from_values([30, 40, 10]),
+                IdMap.from_values(["PAD", 15, 11, 12, 17, 14, 13]),
+                Interactions(
+                    pd.DataFrame(
+                        [
+                            [0, 1, 1.0, "2021-11-25"],
+                            [1, 2, 1.0, "2021-11-25"],
+                            [0, 3, 2.0, "2021-11-26"],
+                            [1, 4, 1.0, "2021-11-26"],
+                            [0, 2, 1.0, "2021-11-27"],
+                            [2, 5, 1.0, "2021-11-28"],
+                            [2, 2, 1.0, "2021-11-29"],
+                            [2, 3, 1.0, "2021-11-29"],
+                            [2, 6, 1.0, "2021-11-30"],
+                        ],
+                        columns=[Columns.User, Columns.Item, Columns.Weight, Columns.Datetime],
+                    ),
+                ),
+            ),
+        ),
+    )
+    def test_process_dataset_train(
+        self,
+        dataset: Dataset,
+        data_preparator: SessionEncoderDataPreparatorBase,
+        expected_interactions: Interactions,
+        expected_item_id_map: IdMap,
+        expected_user_id_map: IdMap,
+    ) -> None:
+        actual = data_preparator.process_dataset_train(dataset)
+        assert_id_map_equal(actual.user_id_map, expected_user_id_map)
+        assert_id_map_equal(actual.item_id_map, expected_item_id_map)
+        assert_interactions_set_equal(actual.interactions, expected_interactions)
+
+    def test_raises_process_dataset_train_when_dense_item_features(
+        self,
+        dataset_dense_item_features: Dataset,
+        data_preparator: SessionEncoderDataPreparatorBase,
+    ) -> None:
+        with pytest.raises(ValueError):
+            data_preparator.process_dataset_train(dataset_dense_item_features)
+
+    @pytest.mark.parametrize(
+        "expected_user_id_map, expected_item_id_map, expected_interactions",
+        (
+            (
+                IdMap.from_values([10, 20]),
+                IdMap.from_values(["PAD", 15, 11, 12, 17, 14, 13]),
+                Interactions(
+                    pd.DataFrame(
+                        [
+                            [0, 6, 1.0, "2021-11-30"],
+                            [0, 2, 1.0, "2021-11-29"],
+                            [0, 3, 1.0, "2021-11-29"],
+                            [0, 5, 1.0, "2021-11-28"],
+                            [1, 6, 9.0, "2021-11-28"],
+                        ],
+                        columns=[Columns.User, Columns.Item, Columns.Weight, Columns.Datetime],
+                    ),
+                ),
+            ),
+        ),
+    )
+    def test_transform_dataset_u2i(
+        self,
+        dataset: Dataset,
+        data_preparator: SessionEncoderDataPreparatorBase,
+        expected_interactions: Interactions,
+        expected_item_id_map: IdMap,
+        expected_user_id_map: IdMap,
+    ) -> None:
+        data_preparator.process_dataset_train(dataset)
+        users = [10, 20]
+        actual = data_preparator.transform_dataset_u2i(dataset, users)
+        assert_id_map_equal(actual.user_id_map, expected_user_id_map)
+        assert_id_map_equal(actual.item_id_map, expected_item_id_map)
+        assert_interactions_set_equal(actual.interactions, expected_interactions)
+
+    @pytest.mark.parametrize(
+        "expected_user_id_map, expected_item_id_map, expected_interactions",
+        (
+            (
+                IdMap.from_values([10, 30, 40, 50, 20]),
+                IdMap.from_values(["PAD", 15, 11, 12, 17, 14, 13]),
+                Interactions(
+                    pd.DataFrame(
+                        [
+                            [0, 6, 1.0, "2021-11-30"],
+                            [0, 2, 1.0, "2021-11-29"],
+                            [0, 3, 1.0, "2021-11-29"],
+                            [1, 2, 1.0, "2021-11-27"],
+                            [1, 3, 2.0, "2021-11-26"],
+                            [1, 1, 1.0, "2021-11-25"],
+                            [2, 2, 1.0, "2021-11-25"],
+                            [2, 4, 1.0, "2021-11-26"],
+                            [0, 5, 1.0, "2021-11-28"],
+                            [4, 6, 9.0, "2021-11-28"],
+                        ],
+                        columns=[Columns.User, Columns.Item, Columns.Weight, Columns.Datetime],
+                    ),
+                ),
+            ),
+        ),
+    )
+    def test_tranform_dataset_i2i(
+        self,
+        dataset: Dataset,
+        data_preparator: SessionEncoderDataPreparatorBase,
+        expected_interactions: Interactions,
+        expected_item_id_map: IdMap,
+        expected_user_id_map: IdMap,
+    ) -> None:
+        data_preparator.process_dataset_train(dataset)
+        actual = data_preparator.transform_dataset_i2i(dataset)
+        assert_id_map_equal(actual.user_id_map, expected_user_id_map)
+        assert_id_map_equal(actual.item_id_map, expected_item_id_map)
+        assert_interactions_set_equal(actual.interactions, expected_interactions)
